% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_null.R
\name{best_null}
\alias{best_null}
\title{Calculates the null model, model with all random terms}
\usage{
best_null(model)
}
\arguments{
\item{model}{any GLMM model}
}
\description{
Calculates the null model, model with all random terms
}
