% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null.R
\name{fit_null}
\alias{fit_null}
\title{Function to fit the Drift model}
\usage{
fit_null(ab, site, region, spp, ...)
}
\arguments{
\item{ab}{vector of species abundance}

\item{site}{vector of localities}

\item{region}{vector of regions (localities should be nested within regions)}

\item{spp}{vector of species identities}

\item{...}{any parameter from glmer}
}
\value{
A linear mixed model of class glmer
}
\description{
Each observation must me relative to a particular species at a particular sampling site.
}
