% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_selection_drift.R
\name{fit_selection_drift}
\alias{fit_selection_drift}
\title{Function to fit the Trait-mediated Selection & Drift model}
\usage{
fit_selection_drift(ab, trait = NULL, grad, site, region, spp, ...)
}
\arguments{
\item{ab}{vector of species abundance}

\item{trait}{vector of species traits. If `NULL` a model with a random effect to correct for the absence of traits is fitted}

\item{grad}{vector of environmental variable}

\item{site}{vector of localities}

\item{region}{vector of regions (localities should be nested within regions)}

\item{spp}{vector of species identities}

\item{...}{any parameter from glmer}
}
\value{
A linear mixed model of class glmer
}
\description{
Each observation must me relative to a particular species at a particular sampling site.
}
