% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_drift.R
\name{r2_drift}
\alias{r2_drift}
\title{Calculates partinioned R-squared for drift model}
\usage{
r2_drift(model, null_model)
}
\arguments{
\item{model}{drift model}

\item{null_model}{the correspondent null model}
}
\description{
Calculates partinioned R-squared for drift model
}
