% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_selection.R
\name{r2_selection}
\alias{r2_selection}
\title{Calculates partinioned R-squared for selection model}
\usage{
r2_selection(model, null.model)
}
\arguments{
\item{model}{selection model}

\item{null.model}{the correspondent null model}
}
\value{

}
\description{
Calculates partinioned R-squared for selection model
}
