% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_selection.R
\name{selection}
\alias{selection}
\title{Function to fit the Selection model}
\usage{
selection(ab, trait = NULL, grad, site, spp, ...)
}
\arguments{
\item{ab}{vector of species abundance}

\item{trait}{vector of species traits. If `NULL` a model with a random effect to correct for the absence of traits is fitted}

\item{grad}{vector of environmental variable}

\item{site}{vector of localities}

\item{spp}{vector of species identities}

\item{...}{any parameter from glmer}
}
\value{
A linear mixed model of class glmer
}
\description{
Each observation must me relative to a particular species at a particular sampling site.
}
